+++ Eve/IrcExceptions.cs
@@ -1,3 +1,0 @@
-﻿namespace Eve.Classes {
-    // custom exceptions will be implemented in future, maybe
-}